net: socket: mgmt: use uintptr_t for the nm_pid field
This may contain a pointer so make sure it is sufficiently wide
on 64-bit targets.

+++ samples/net/sockets/net_mgmt/src/main.c
@@ -97,7 +97,7 @@ static void listener(void)
 
 	sockaddr.nm_family = AF_NET_MGMT;
 	sockaddr.nm_ifindex = 0; /* Any network interface */
-	sockaddr.nm_pid = (int)k_current_get();
+	sockaddr.nm_pid = (uintptr_t)k_current_get();
 	sockaddr.nm_mask = NET_EVENT_IPV6_DAD_SUCCEED |
 			    NET_EVENT_IPV6_ADDR_ADD |
 			    NET_EVENT_IPV6_ADDR_DEL;
